requirements.txt
tiktoken
chromadb
langchain
flask
GitPython
python-dotenv
google-generativeai
langchain_community
ipykernel
langchain-google-genai